pyproject.toml
[tool.poetry]
name = "bird-species-recognition-aml2024"
version = "0.1.0"
description = "Deep learning-based classification of 200 bird species using image recognition techniques"
authors = ["Szymon Budziak"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
torch = "^2.5.1"
torchvision = "^0.20.1"
pyyaml = "^6.0.2"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
notebook = "^7.2.2"
black = "^24.10.0"
mypy = "^1.13.0"
kaggle = "^1.6.17"
types-pyyaml = "^6.0.12.20240917"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
